Split folded booking tests by concern
Move the processParentApiBooking DB-flow tests into

folded-booking/parent-booking.test.ts and keep the pure parse/apply

tests in folded-booking.test.ts, so each file stays under the

~400-line target. Both own the source via the mirror/folder mapping.
